+++ console/migrations/m220223_122342_add_column_qp_client_selling_to_quote_price_tbl.php
@@ -1,5 +1,6 @@
 <?php
 
+use src\helpers\app\DBHelper;
 use yii\db\Migration;
 
 /**
@@ -12,7 +13,9 @@ class m220223_122342_add_column_qp_client_selling_to_quote_price_tbl extends Mig
      */
     public function safeUp()
     {
-        $this->addColumn('{{%quote_price}}', 'qp_client_selling', $this->decimal(10, 2));
+        if (!DBHelper::isColumnExist('quote_price', 'qp_client_selling')) {
+            $this->addColumn('{{%quote_price}}', 'qp_client_selling', $this->decimal(10, 2));
+        }
     }
 
     /**
@@ -20,6 +23,8 @@ class m220223_122342_add_column_qp_client_selling_to_quote_price_tbl extends Mig
      */
     public function safeDown()
     {
-        $this->dropColumn('{{%quote_price}}', 'qp_client_selling');
+        if (DBHelper::isColumnExist('quote_price', 'qp_client_selling')) {
+            $this->dropColumn('{{%quote_price}}', 'qp_client_selling');
+        }
     }
 }
